[clangBasic] Emit SARIF Diagnostics through `clang::SarifDocumentWriter`
Create an interface for writing SARIF documents from within `clang`.

The primary intent of this change is to introduce the interface
`clang::SarifDocumentWriter`, which allows incrementally adding diagnostic
properties to a JSON backed document.

The proposed interface is not yet connected to the compiler internals, which
will be covered in future work. As such this change will not change the
input/output interface of `clang`.

- Create class `clang::SarifDocumentWriter` that provides an interface to writing SARIF documents
- Create other value semantic types that can be used to represent
  objects in the SARIF specification

The implementation targets SARIF 2.1.0

Differential Revision: https://reviews.llvm.org/D109701